Make CLI teardown regression coverage deterministic
Cancel synchronously at the fake project boundary and share the signaling time provider between command tests.
